+++ CHANGELOG.md
@@ -20,3 +20,4 @@ Created on Dec 20, 2023
 - Beta Release: Added more transaction options and minor improvements on December 25, 2023 -> **0.46**
 - Beta Release: Transaction deletion and system improvement on December 25, 2023 -> **0.48**
 - Beta Release: Transaction DB model adjustment + deletion UI adjustment on December 26, 2023 -> **0.49**
+- Beta Release: Transaction DB model adjustment + add migration on December 26, 2023 -> **0.491**